auth_microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
